# docs/swagger.yaml
definitions:
  dto.TrojanItemOutput:
    properties:
      asset_id:
        type: string
      is_Conn:
        type: integer
    type: object
  dto.TrojanListOutput:
    properties:
      list:
        description: 列表
        items:
          $ref: '#/definitions/dto.TrojanItemOutput'
        type: array
      total:
        description: 总数
        type: integer
    type: object
  middleware.Response:
    properties:
      data:
        type: object
      errmsg:
        type: string
      errno:
        type: integer
      stack:
        type: object
      trace_id:
        type: object
    type: object
info:
  contact: {}
  license: {}
paths:
  /demo/info:
    get:
      consumes:
      - application/json
      description: 模板简介
      operationId: /demo/info
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: 模板-Demo
      tags:
      - 模板
  /trojan/echo:
    get:
      consumes:
      - application/json
      description: 木马通信
      operationId: /trojan/echo
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  type: string
              type: object
      summary: WS即时通
      tags:
      - 浮标管理
  /trojan/service/list:
    get:
      consumes:
      - application/json
      description: 获取在线木马服务列表
      operationId: /trojan/service/list
      parameters:
      - description: 关键词
        in: query
        name: info
        type: string
      - description: 每页个数
        in: query
        name: limit
        required: true
        type: integer
      - description: 当前页数
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/middleware.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.TrojanListOutput'
              type: object
      summary: 获取在线木马服务列表
      tags:
      - 浮标管理
swagger: "2.0"